Refactor Profile
