+++ app/src/main/res/layout/activity_profile_settings.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context="org.wit.archaeologicalfieldwork.activities.user.ProfileSettingsActivity">
+  tools:context="org.wit.archaeologicalfieldwork.views.user.settings.SettingsView">
 
   <RelativeLayout
     android:layout_width="match_parent"
