+++ app/src/main/res/layout/card_hillfort.xml
@@ -9,7 +9,8 @@
   <RelativeLayout
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:padding="16dp">
+    android:padding="16dp"
+    android:background="@color/colorPrimary">
 
     <ImageView
       android:id="@+id/imageIcon"
