+++ app/src/main/res/menu/menu_profile.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto">
+
+  <item
+    android:id="@+id/item_edit"
+    android:title="@string/menu_profile"
+    android:icon="@drawable/baseline_edit_white_18dp"
+    app:showAsAction="always"/>
+</menu>
